+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - ONTOLOGY-20 Changed free-text selection styling logic
 
 ### Deprecated
+- ONTOLOGY-20 Removed versioning
 
 ### Removed